+++ test/leiningen/clojure_lsp_test.clj
@@ -1,6 +1,7 @@
 (ns leiningen.clojure-lsp-test
-  (:require [clojure.test :refer :all]
-            [leiningen.clojure-lsp :as clojure-lsp]))
+  (:require
+   [clojure.test :refer :all]
+   [leiningen.clojure-lsp :as clojure-lsp]))
 
 (deftest args-test
   (testing "Should return the arguments when project settings is an empty map"
@@ -13,7 +14,7 @@
 
   (testing "Should return the arguments when there is project settings"
     (is (= ["diagnostics" "--settings" "{:foo 1}"]
-           (clojure-lsp/args ["diagnostics"] {:foo 1}))))
+           (clojure-lsp/args ["diagnostics"] {:settings {:foo 1}}))))
 
   (testing "Should return the arguments when there is settings options and also project settings"
     (is (= ["diagnostics" "--settings" "{:foo 1, :bar \"Test\"}"]