+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix --settings flag
+- Replace clojure-lsp JVM with native binary for way better performance.
 
 ## 1.4.21